Transfer initialSate to different file, import it in book.js
